UGRID-14, UGRID-16, UGRID-17, UGRID-19: Pitchfork layout (#21)

+++ scripts/docker/main.sh
@@ -3,17 +3,46 @@
 set -e
 
 source /workspace/scripts/docker/error.sh
+source /workspace/scripts/docker/utilities.sh
 
 export netCDFCxx_DIR=${THIRD_PARTY_INSTALL_DIR}/netcdf_cxx4/lib64/cmake/netCDF/
 
 build_dir="/workspace/build"
-if ${CLEAN_BUILD}; then
+
+if [[ "${CLEAN_BUILD}" = "true" && -d ${build_dir} ]]; then
+  print_text_box "Clean build"
+  # cmake --build ${build_dir} --target clean ||
+  #   error "[cmake] Failed to clean project"
   rm -fr ${build_dir}
 fi
 
-cmake -S . -B ${build_dir} -DCMAKE_BUILD_TYPE=Release -DCMAKE_PREFIX_PATH="${THIRD_PARTY_INSTALL_DIR}" ||
+# configure
+print_text_box "Configure build"
+cmake \
+  -S . \
+  -B ${build_dir} \
+  -DCMAKE_BUILD_TYPE=${BUILD_TYPE} \
+  -DCMAKE_PREFIX_PATH="${THIRD_PARTY_INSTALL_DIR}" ||
   error "[cmake] Failed to configure project"
-cmake --build ${build_dir} --config Release --parallel ||
+
+verbose_switch=""
+if [ "${VERBOSE_BUILD}" = "true" ]; then
+  verbose_switch="--verbose"
+fi
+
+# build
+print_text_box "Build"
+cmake \
+  --build ${build_dir} \
+  --config ${BUILD_TYPE} \
+  --parallel \
+  ${verbose_switch} ||
   error "[cmake] Failed to build project"
 
-${build_dir}/bin/UGridApiTests
+# run tests
+print_text_box "Test"
+ctest --output-on-failure --extra-verbose --test-dir ${build_dir}
+
+# install
+print_text_box "Install"
+cmake --install ${build_dir} --prefix ${build_dir}/install
